Move camera.ts, files.ts and socket.ts into the common API
