+++ client/src/api/common/socket.ts
