+++ client/src/api/v2/sensors/data.ts
@@ -1,5 +1,5 @@
 import { useState, useCallback } from 'react';
-import { useChannel } from '../socket';
+import { useChannel } from '../../common/socket';
 
 /* eslint-disable camelcase */
 export interface SensorData {
