+++ client/src/api/v2/settings.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect, useCallback } from 'react';
-import { useChannel, emit } from './socket';
+import { useChannel, emit } from '../common/socket';
 
 export interface DASSettings {
   /** Whether to publish data to online broker */
