+++ client/src/components/common/camera_settings/OverlayMessageContainer.tsx
@@ -1,4 +1,4 @@
-import { sendMessage } from 'api/v2/camera';
+import { sendMessage } from 'api/common/camera';
 import React from 'react';
 import OverlayMessage from 'components/common/camera_settings/OverlayMessage';
