+++ client/src/components/common/camera_settings/OverlaySelectionContainer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useCameraConfig } from 'api/v2/camera';
+import { useCameraConfig } from 'api/common/camera';
 import OverlaySelection from 'components/common/camera_settings/OverlaySelection';
 
 /**
